llm_service/requirements.txt
python-dotenv==1.0.0
pika==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
ollama==0.1.6